Renaming folder & controller
